requirements.txt
docopt==0.6.2
fpdf2==2.8.2
